Реализовал работу с базой на получение

+++ app/src/main/java/com/client/googlenotes/data/database/models/NoteTable.kt
@@ -8,5 +8,5 @@ import io.requery.Persistable
 interface NoteTable: Persistable {
 
     @Key
-    fun getId(): Int
+    fun getId(): Long
 }
